test(IDX): drop dependency on //rs/tests from //rs/tests/testing_verification:...
