+++ rs/tests/driver/BUILD.bazel
@@ -1,5 +1,5 @@
 load("@bazel_skylib//rules:common_settings.bzl", "bool_flag")
-load("@rules_rust//rust:defs.bzl", "rust_library")
+load("@rules_rust//rust:defs.bzl", "rust_library", "rust_test")
 
 package(default_visibility = ["//rs:system-tests-pkg"])
 
@@ -178,4 +178,9 @@ rust_library(
         "@crate_index//:wat",
         "@crate_index//:zstd",
     ],
+)
+
+rust_test(
+    name = "generic_workload_engine",
+    crate = ":ic-system-test-driver",
 )
