+++ rs/tests/src/lib.rs
@@ -6,9 +6,7 @@ pub mod crypto;
 pub mod custom_domains_integration;
 pub mod execution;
 pub mod icrc1_agent_test;
-pub mod ii_delegation_test;
 pub mod ledger_tests;
-pub mod mainnet_tests;
 pub mod message_routing;
 pub mod networking;
 pub mod nns_tests;
