--- rs/tests/src/mainnet_tests/mod.rs
@@ -1,1 +1,0 @@
-pub mod mainnet_api;
