+++ rs/tests/testing_verification/Cargo.toml
@@ -12,7 +12,6 @@ ic-registry-subnet-type = { path = "../../registry/subnet_type" }
 ic-system-test-driver = { path = "../driver" }
 slog = { workspace = true }
 spec-compliance = { path = "./spec_compliance" }
-tests = { path = ".." }
 
 [[bin]]
 name = "ic-systest-basic-health-test"
