+++ rs/tests/testing_verification/replicable_mock_test.rs
@@ -1,7 +1,4 @@
-#[rustfmt::skip]
-
 use anyhow::Result;
-
 use ic_system_test_driver::driver::group::SystemTestGroup;
 use ic_system_test_driver::driver::test_env::TestEnv;
 use ic_system_test_driver::systest;
